pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "grandalf>=0.8",
    "huggingface-hub>=0.35.0",
    "ipykernel>=7.2.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.33",
    "langgraph>=1.0.1",
    "numpy>=2.3.3",
    "openai>=1.107.3",
    "pypdf>=6.1.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.0",
    "transformers>=4.56.1",
]
